pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "galvanize-agents"
version = "0.1.0"
description = "Multi-agent system for Galvanize NMTC/HTC audience intelligence and content strategy"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "Galvanize Sustainable Capital" }]
dependencies = [
    "crewai>=0.80.0,<1.10.0",
    "crewai-tools>=0.14.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
galvanize = "galvanize.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
